refactor: 핸들러를 동적으로 생성하는 부분 책임 분리
